fix: GameBoard: moved RandInt into TU's anon namespace
