+++ src/headers/gameboard.hpp
@@ -3,28 +3,10 @@
 
 #include "point2d.hpp"
 #include "tile.hpp"
-#include <chrono>
-#include <random>
 #include <tuple>
+#include <vector>
 
 namespace Game {
-
-class RandInt {
-public:
-  using clock = std::chrono::system_clock;
-  RandInt() : dist{0, std::numeric_limits<int>::max()} {
-    seed(clock::now().time_since_epoch().count());
-  }
-  RandInt(const int low, const int high) : dist{low, high} {
-    seed(clock::now().time_since_epoch().count());
-  }
-  int operator()() { return dist(re); }
-  void seed(const unsigned int s) { re.seed(s); }
-
-private:
-  std::minstd_rand re;
-  std::uniform_int_distribution<> dist;
-};
 
 using tile_data_array_t = std::vector<Tile>;
